requirements.txt
# Core libraries
numpy==1.26.3
torch==2.0.1          
torchvision==0.15.2  
opencv-python==4.8.0.74
Pillow==9.4.0
pandas==2.0.3
matplotlib==3.7.1
scikit-learn==1.2.2
seaborn==0.12.2
imgaug==0.4.0
torchinfo==1.8.0
wandb==0.15.2

# Optional: Other useful packages
tqdm==4.65.0
glob2==0.7